Remove `rebind()` method and add  `bind()` method to actually bind `DatagramSocket` to the device network.
Previous implementation only reinitialize the `DatagramSocket` but not binding to any network. This change is necessary because we need internet connection to access Google Map. So, instead of binding all app process to device network, we should bind the `Socket` that does need to communicate with the device (UDP and WebSocket).